+++ packages/orq-rc/RELEASES.md
@@ -2809,3 +2809,13 @@ Based on:
 - [typescript v3.10.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.39 - packages/orq-rc
+
+## 2025-08-02 06:22:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.596.0 (2.672.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.40 - packages/orq-rc